Tutorial using a new model - necessary changes

+++ src/reconstruction/modelGeneration/massBalance/checkMassChargeBalance.m
@@ -2,7 +2,7 @@ function [massImbalance, imBalancedMass, imBalancedCharge, imBalancedRxnBool, El
 % Tests for a list of reactions if these reactions are
 % mass-balanced by adding all elements on left hand side and comparing them
 % with the sums of elements on the right hand side of the reaction.
-% A reaction is considered elementally imbalanced if any of the molecular 
+% A reaction is considered elementally imbalanced if any of the molecular
 % species involved is missing a chemical formula.
 %
 % USAGE:
@@ -88,7 +88,7 @@ end
 
 E=full(E);
 
-% A reaction is considered elementally imbalanced if any of the molecular 
+% A reaction is considered elementally imbalanced if any of the molecular
 % species involved is missing a chemical formula.
 imBalancedRxnBool=any(massImbalance, 2) | any(model.S(missingFormulaeBool, :))';
 
@@ -190,7 +190,7 @@ if isfield(model, 'metCharges')
         end
     end
 
-    imBalancedCharge=model.S'*model.metCharges;
+    imBalancedCharge=model.S' * double(model.metCharges);
 end
 
 if printLevel==-1 && isfield(model,'SIntRxnBool')
